Dood aan de float!

+++ src/TLSampleMethodResistive.h
@@ -37,7 +37,7 @@ struct TLStructSampleMethodResistive {
 	int pin;
 	int gndPin;
 	int useInternalPullup;
-	float valueMax;
+	int32_t valueMax;
 };
 
 int TLSampleMethodResistivePreSample(struct TLStruct * data, uint8_t nSensors,
